SourceLevelDebugPkg: DebugAgentLib: Fix clang/Xcode 5 compile/link errors
Move ExceptionStubHeaderSize from 16 to 32 bits to work around clang relocation limitation. Use movw, not move for 32-bit segment register operations. 

Contributed-under: TianoCore Contribution Agreement 1.0




git-svn-id: https://svn.code.sf.net/p/edk2/code/trunk/edk2@16067 6f19259b-4bc3-4df7-8a09-765794883524

+++ SourceLevelDebugPkg/Library/DebugAgent/DebugAgentCommon/DebugAgent.h
@@ -67,7 +67,7 @@
 
 extern UINTN  Exception0Handle;
 extern UINTN  TimerInterruptHandle;
-extern UINT16 ExceptionStubHeaderSize;
+extern UINT32 ExceptionStubHeaderSize;
 extern BOOLEAN mSkipBreakpoint;
 extern EFI_VECTOR_HANDOFF_INFO mVectorHandoffInfoDebugAgent[];
 extern UINTN                   mVectorHandoffInfoCount;
